+++ tests/testthat/test_decode.R
@@ -14,20 +14,3 @@ test_that(
   }
 )
 
-test_that(
-  "decode one structural attribute",
-  {
-    dt <- decode("GERMAPARLMINI", sAttribute = "date")
-    expect_equal(length(unique(dt[["date"]])), 5)
-    expect_equal(sum(dt[["cpos_left"]]), 540378099)
-  }
-)
-
-test_that(
-  "decode several structural attributes",
-  {
-    dt <- decode("GERMAPARLMINI", sAttribute = c("date", "speaker"))
-    expect_equal(ncol(dt), 4)
-    expect_equal(length(unique(dt[["speaker"]])), 124)
-  }
-)